\name{minque-package}
\alias{minque-package}
\alias{minque}
\docType{package}
\title{
An R Package for Various Linear Mixed Models Analyses

}
\description{
The package can be used for various linear mixed linear model analyses. 
}
\details{
\tabular{ll}{
Package: \tab MINQUE\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-09-06\cr
License: \tab GPL-2\cr
}
 An overview of how to use the package, including the most important functions
}
\author{
Jixiang Wu, Assistant Professor of Quantitative Genetics/Biostatistics,
\\Jixiang Wu <qgtools@gmail.com>

}
\references{

Miller, R. G. 1974. The jackknife - a review. Biometrika, 61:1- 15.

Patterson, H. D. and Thompson, R. 1971. Recovery of inter-block information when block sizes are unequal. Biometrika, 58: 545-554.

Rao, C.R. 1971. Estimation of variance and covariance components-MINQUE theory. J Multiva Ana 1:19

Rao, C. R. and Kleffe, J. 1980. Estimation of variance components. In Handbook of Statistics. Vol. l: 1-40. Krishnaiah, P. R. ed. New York. North-Holland.

Searle, S. R., Casella, G. and McCulloch, C. E. 1992. Variance Components. John Wiley & Sons, Inc. New York.

Wu J (2012) GenMod: An R package for various agricultural data analyses.  ASA, CSSA, and SSSA 2012 International Annual Meetings, Cincinnati, OH, p 127

Wu J., Bondalapati K., Glover K., Berzonsky W., Jenkins J.N., McCarty J.C. 2013. Genetic analysis without replications: model evaluation and application in spring wheat. Euphytica. 190:447-458 

Zhu J. 1989. Estimation of Genetic Variance Components in the General Mixed Model. Ph.D. Dissertation, NC State University, Raleigh, U.S.A

Zhu J. 1993. Methods of predicting genotype value and heterosis for offspring of hybrids. (Chinese). Journal of Biomathematics, 8(1): 32-44
}

\keyword{ MINQUE }
\keyword{ Linear mixed model }
\keyword{ Jackknife }
\keyword{ REML }

